requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.1
plyer==2.1.0
psutil==5.9.8
python-dotenv==1.0.0
sendgrid==6.11.0